All requirement completed
REQUIREMENTS:
As a customer I wish to order a drink/snack.
As a customer I wish to see what I have ordered.
As a customer I wish to add to my current order for a drink/snack.
As a customer I wish to cancel my order for a drink/snack.
As the admin I wish to view a customer’s order(s).
As the admin I wish to read the details of the drinks/snacks I have for sale.
As the admin I wish to edit the details of the drinks/snacks I have for sale.
As the admin I wish to enter(create) details of the drinks/snacks I have for sale.
As the admin I wish to withdraw(disable) a drink/snack from sale.